requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn
h5py==2.10.0
tensorflow==1.15
tensorflow-gpu==1.15.0
keras==2.1.6
clint
crontab
tablib

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
fastapi
uvicorn
requests
opencv-python
pydeck
streamlit

# utilities
six>=1.14
joblib
memoized-property
opencv-python

# for coco-viewer
scikit-image
pycocotools
Pillow==9.0.0

# geo packages
geopandas
shapely
geopy

#cloud
google.cloud
